requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
python-multipart
requests
groq
python-dotenv
fastapi-cors
transformers
